requirements.txt
PyPDF2==2.4
pdf2image
Pillow